Changed dice chat creation behaviour
