+++ apps/stygian_web/lib/stygian_web/live/chat_live/chat_live.html.heex
@@ -33,6 +33,8 @@
         id={@current_character.id}
         current_character={@current_character}
         map={@map}
+        attributes={@attributes}
+        skills={@skills}
       />
     </.modal>
